heroku fix bug

+++ src/main/java/com/boots/controller/AdminController.java
@@ -24,7 +24,7 @@ public class AdminController {
     public String deleteUser(@RequestParam(required = true, defaultValue = "") Long userId,
                              @RequestParam(required = true, defaultValue = "") String action,
                              Model model) {
-        if (action.equals("delete")) {
+        if (action.equals("delete") && userId != 1) {
             userService.deleteUser(userId);
         }
         return "redirect:/admin";
